+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to
 - Claim an order generated from a batch order through voucher code
 - Add description to order group model
 - Add discounted price to course product relation serializer
+- Add description to order group form in the back office
 
 ### Changed